refactor: move GameService to dedicated service package for better code hygiene
