+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/TestServiceFactory.kt
@@ -1,7 +1,7 @@
 package at.aau.hexabrawl.websocketserver
 
 import at.aau.hexabrawl.websocketserver.service.CombatService
-import at.aau.hexabrawl.websocketserver.model.GameService
+import at.aau.hexabrawl.websocketserver.service.GameService
 import at.aau.hexabrawl.websocketserver.service.ConnectivityService
