+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/controller/CheatGiftBlockTest.kt
@@ -8,6 +8,7 @@ import org.mockito.Mockito.*
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import at.aau.hexabrawl.websocketserver.TestServiceFactory
+import at.aau.hexabrawl.websocketserver.service.GameService
 
 /**
  * Tests, dass /move, /end-turn, /buy-farm und /buy-unit blockiert sind
