+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/model/DisconnectCleanupServiceTest.kt
@@ -8,6 +8,7 @@ import org.mockito.Mockito.*
 import org.springframework.messaging.simp.SimpMessagingTemplate
 import at.aau.hexabrawl.websocketserver.TestServiceFactory
 import at.aau.hexabrawl.websocketserver.service.DisconnectCleanupService
+import at.aau.hexabrawl.websocketserver.service.GameService
 
 /**
  * Tests fuer den DisconnectCleanupService.
