+++ src/test/kotlin/at/aau/hexabrawl/websocketserver/model/FieldConquestTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import at.aau.hexabrawl.websocketserver.TestServiceFactory
+import at.aau.hexabrawl.websocketserver.service.GameService
 
 
 class FieldConquestTest {
